+++ R/ui_helpers.R
@@ -86,7 +86,7 @@ app_brand_title <- function(version) {
 }
 
 app_stylesheet_link <- function(version) {
-  tags$link(rel = "stylesheet", type = "text/css", href = paste0("style.css?v=", version, "-save-layout-1"))
+  tags$link(rel = "stylesheet", type = "text/css", href = paste0("style.css?v=", version, "-hint8-layout-1"))
 }
 
 app_script_link <- function(version) {